update test for domain class
