+++ domain/src/main/kotlin/fr/sacane/jmanager/domain/models/User.kt
@@ -9,13 +9,19 @@ class UserId(private val id: Long){
 
 
 class Password(val value: String){
-
+    init{
+        require(value.isNotBlank() && value.isNotEmpty()){
+            "Given password is blank or empty"
+        }
+    }
 
     fun get(): ByteArray{
         return Hash.hash(value)
     }
 
-
+    fun matchWith(password2: Password): Boolean {
+        return get().contentEquals(password2.get())
+    }
 }
 
 class User(
