+++ domain/src/main/kotlin/fr/sacane/jmanager/domain/port/serverside/UserDirectory.kt
@@ -7,7 +7,7 @@ import fr.sacane.jmanager.domain.models.*
 interface UserTransaction {
     fun findById(userId: UserId): Ticket?
     fun checkUser(pseudonym: String, pwd: Password): Ticket?
-    fun findByPseudonym(pseudonym: String): Ticket?
+    fun findByPseudonym(pseudonym: String): User?
     fun create(user: User): User?
     fun save(user: User): User?
     fun getUserToken(userId: UserId): Token?
